# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: SpaceOS Scheduling API
  version: 1.0.0-preview.1
  summary: Read-only scheduling contract — plans, resources and operation standards.
  description: |
    Every response is tenant-scoped by the caller's token. The tenant is never taken from a
    request header, and a caller whose tenant has not enabled `spaceos.scheduling` is refused
    before any data is read.

    Timestamps are ISO-8601 UTC strings and time zones are IANA identifiers: no platform-side
    library choice appears on the wire (ADR-070 D2).
servers:
  - url: /api/scheduling/v1
security:
  - bearerAuth: []
...
paths:
  /runs:
    get:
      operationId: listScheduleRuns
      summary: Lists scheduling runs, newest first.
      tags: [scheduling]
      parameters:
        - name: projectId
          in: query
          required: false
          description: Restricts the result to one Kernel project.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The runs visible to the caller's tenant.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleRunSummary'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /runs/{runId}:
    get:
      operationId: getScheduleRun
      summary: Gets one scheduling run and the state of its revision chain.
      tags: [scheduling]
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: The run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleRunSummary'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /runs/{runId}/proposal:
    get:
      operationId: getProposal
      summary: Gets the active plan of a run.
      description: |
        Returns the published revision when there is one, otherwise the newest revision that
        is still live. A discarded or superseded revision is never returned.
      tags: [scheduling]
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: The plan in force, with the provenance needed to act on it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /resources:
    get:
      operationId: listResources
      summary: Lists resources with their current calendar revision.
      tags: [scheduling]
      responses:
        '200':
          description: One entry per resource that has an open calendar revision.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceCalendar'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /resources/{resourceKey}/calendar:
    get:
      operationId: getResourceCalendar
      summary: Gets a resource's current calendar revision in full.
      tags: [scheduling]
      parameters:
        - $ref: '#/components/parameters/ResourceKey'
      responses:
        '200':
          description: The recurring pattern and its dated exceptions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceCalendarDetail'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /resources/{resourceKey}/overload:
    get:
      operationId: getResourceOverload
      summary: Lists periods where committed demand exceeds the resource's capacity.
      description: |
        A resource with no calendar revision answers 404 rather than an empty list: without a
        calendar there is no capacity to measure demand against.
      tags: [scheduling]
      parameters:
        - $ref: '#/components/parameters/ResourceKey'
      responses:
        '200':
          description: Contiguous oversubscribed periods, each reported with its peak demand.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OverloadSpan'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /standards:
    get:
      operationId: listOperationStandards
      summary: Lists operation standards and the state of their revision chains.
      tags: [scheduling]
      responses:
        '200':
          description: The standards visible to the caller's tenant.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OperationStandardSummary'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /standards/{sourceTaskKey}/revisions:
    get:
      operationId: listStandardRevisions
      summary: Lists every revision of one operation standard.
      tags: [scheduling]
      parameters:
        - name: sourceTaskKey
          in: path
          required: true
          description: Stable key from the source catalogue.
          schema:
            type: string
      responses:
        '200':
          description: Every revision, quarantined ones included.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StandardRevision'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued access token carrying the tenant and its enabled modules.
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      description: Scheduling run identifier.
      schema:
        type: string
        format: uuid
    ResourceKey:
      name: resourceKey
      in: path
      required: true
      description: Stable resource key.
      schema:
        type: string
  responses:
    Unauthorized:
      description: No valid token was presented.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: |
        The token is valid but the tenant has not enabled `spaceos.scheduling`, or a tenant
        header was sent that the token does not authorise.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: |
        No such resource for this tenant. Another tenant's data answers 404 as well — a 403
        would confirm that the identifier exists.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      description: RFC 9457 problem document with the platform-wide correlation identifier.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        correlationId:
          type: string
          description: Same identifier across every layer that handled the request.
      required: [title, status]
...
    ScheduleRunSummary:
      type: object
      properties:
        runId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        createdAtUtc: { type: string, format: date-time }
        publishedRevisionHash:
          type: [string, 'null']
          description: Content hash of the plan in force; null when nothing is published.
        revisionCount: { type: integer }
      required: [runId, projectId, createdAtUtc, publishedRevisionHash, revisionCount]
    OperationPlan:
      type: object
      properties:
        operationId: { type: string }
        scope: { $ref: '#/components/schemas/WorkScope' }
        resourceKey: { type: string }
        startMinute: { type: number }
        finishMinute: { type: number }
        automaticallyPlanned:
          type: boolean
          description: False when the operation is visible but was not placed automatically.
        standardRevision:
          type: [integer, 'null']
          description: Revision of the norm the times were derived from.
        sourceRevisions:
          type: object
          additionalProperties: { type: string }
          description: |
            The consumer's own provenance, reflected back unchanged. The platform stores these
            values and never resolves them.
        startUtc:
          type: [string, 'null']
          format: date-time
          description: |
            When the operation starts, resolved against the calendar revisions this plan is
            PINNED to (see calendarRevisions on the proposal) — not against today's calendars.
            The same revision therefore always answers with the same instants, including after a
            newer calendar revision is approved.

            Null on revisions computed before the plan recorded its timeline origin: those
            cannot be dated, and a plausible-but-invented date would be worse than none.
        finishUtc:
          type: [string, 'null']
          format: date-time
          description: |
            When the operation's WORKING time is complete. Because durations are working time,
            this can be days after startUtc for work that spans shifts, nights or a weekend.
      required:
        - operationId
        - scope
        - resourceKey
        - startMinute
        - finishMinute
        - automaticallyPlanned
        - standardRevision
        - sourceRevisions
        - startUtc
        - finishUtc
    DependencyEdge:
      type: object
      properties:
        predecessorOperationId: { type: string }
        successorOperationId: { type: string }
        relationCode:
          type: string
          enum: [FS, SS, FF, SF]
        lagMinutes:
          type: number
          description: Positive for a lag, negative for a lead.
        earliestStartMinute:
          type: [number, 'null']
        startSource:
          type: [string, 'null']
          enum: [fixed_override, partial_release, dependency, null]
          description: Which rule produced the resolved start bound.
        warnings:
          type: array
          items:
            type: string
            enum: [partial_release_delays_fs_start]
        releaseThresholdFraction:
          type: [number, 'null']
          description: >-
            Fraction of the predecessor's working time after which the successor was released,
            when the edge carries a partial release; null otherwise. "Released at 0.8" is a
            different agreement from "at 0.5" even when the resulting dates coincide.
        lagKind:
          type: string
          enum: [working, elapsed]
          default: working
          description: >-
            What lagMinutes is measured in. "working" counts working time on the successor's
            calendar and pauses when the resource does; "elapsed" is real clock time — a curing
            or drying period runs through the weekend. Absent on older payloads, where it always
            meant "working".
      required:
        - predecessorOperationId
        - successorOperationId
        - relationCode
        - lagMinutes
        - earliestStartMinute
        - startSource
        - warnings
        - releaseThresholdFraction
        - lagKind
    Proposal:
      type: object
      description: The plan in force, with everything needed to reproduce and quote it.
      properties:
        runId: { type: string, format: uuid }
        revisionHash:
          type: string
          description: Quote this back when acting on the proposal.
        state:
          type: string
          enum: [proposal, shadow, published, superseded, discarded]
        calculatedAtUtc: { type: string, format: date-time }
        operations:
          type: array
          items: { $ref: '#/components/schemas/OperationPlan' }
        dependencies:
          type: array
          items: { $ref: '#/components/schemas/DependencyEdge' }
        calendarRevisions:
          type: object
          additionalProperties: { type: integer }
          description: Calendar revision per resource; the plan is only reproducible with it.
        ruleSetLabel:
          type: string
          description: Provenance of the rule set in force.
        capacityConflicts:
          type: array
          items: { $ref: '#/components/schemas/CapacityConflict' }
          description: |
            Where this plan would collide with work that is ALREADY COMMITTED on its resources.
            The plan's own operations always respect capacity — the scheduler saw to that — so
            this reports the collision the scheduler could not see: reservations made by other
            runs and other work.

            Computed by the same detector as /resources/{key}/overload, so the two views cannot
            disagree about the same afternoon. Empty when the plan fits beside committed work;
            also empty when the plan carries no dates (see startUtc), because an undated plan
            cannot be compared with anything on the clock.
      required:
        - runId
        - revisionHash
        - state
        - calculatedAtUtc
        - operations
        - dependencies
        - calendarRevisions
        - ruleSetLabel
        - capacityConflicts
    CapacityConflict:
      type: object
      description: A period where the plan plus committed work exceeds a resource's capacity.
      properties:
        resourceKey: { type: string }
        startUtc: { type: string, format: date-time }
        endUtc:
          type: string
          format: date-time
          description: Exclusive end of the conflicting period.
        peakDemand:
          type: number
          description: Highest simultaneous demand inside the period.
        capacity:
          type: number
          description: Capacity the demand was measured against, from the pinned calendar revision.
        peakExcess:
          type: number
          description: peakDemand minus capacity — how bad it gets at worst.
      required:
        - resourceKey
        - startUtc
        - endUtc
        - peakDemand
        - capacity
        - peakExcess
    DayRange:
      type: object
      description: Minutes from local midnight; start inclusive, end exclusive.
      properties:
        startMinuteOfDay: { type: integer }
        endMinuteOfDay: { type: integer }
      required: [startMinuteOfDay, endMinuteOfDay]
    RecurringShift:
      type: object
      properties:
        isoWeekday:
          type: integer
          description: ISO-8601 weekday, Monday = 1.
        shift: { $ref: '#/components/schemas/DayRange' }
        breaks:
          type: array
          items: { $ref: '#/components/schemas/DayRange' }
      required: [isoWeekday, shift, breaks]
...
    ResourceCalendar:
      type: object
      properties:
        resourceKey: { type: string }
        timeZoneId:
          type: string
          description: IANA identifier, for example Europe/Budapest.
        capacity: { type: number }
        capacityPolicy:
          type: string
          enum: [integer, fractional_fte]
        revision: { type: integer }
        isApproved:
          type: boolean
          description: Only an approved revision may be scheduled against.
      required: [resourceKey, timeZoneId, capacity, capacityPolicy, revision, isApproved]
    ResourceCalendarDetail:
      type: object
      properties:
        resourceKey: { type: string }
        revision: { type: integer }
        timeZoneId: { type: string }
        capacity: { type: number }
        capacityPolicy:
          type: string
          enum: [integer, fractional_fte]
        isApproved: { type: boolean }
        effectiveFromUtc: { type: string, format: date-time }
        effectiveToUtc: { type: [string, 'null'], format: date-time }
        shifts:
          type: array
          items: { $ref: '#/components/schemas/RecurringShift' }
        exceptions:
          type: array
          items: { $ref: '#/components/schemas/CalendarException' }
      required:
        - resourceKey
        - revision
        - timeZoneId
        - capacity
        - capacityPolicy
        - isApproved
        - effectiveFromUtc
        - effectiveToUtc
        - shifts
        - exceptions
    OverloadSpan:
      type: object
      properties:
        startUtc: { type: string, format: date-time }
        endUtc: { type: string, format: date-time }
        peakDemand: { type: number }
        capacity: { type: number }
        peakExcess:
          type: number
          description: Peak demand minus capacity.
      required: [startUtc, endUtc, peakDemand, capacity, peakExcess]
    OperationStandardSummary:
      type: object
      properties:
        sourceTaskKey: { type: string }
        qualifiers:
          type: object
          additionalProperties: { type: string }
        acceptedRevision:
          type: [integer, 'null']
          description: Revision currently usable for planning; null when none is.
        revisionCount: { type: integer }
      required: [sourceTaskKey, qualifiers, acceptedRevision, revisionCount]
    StandardRevision:
      type: object
      properties:
        sourceTaskKey: { type: string }
        revision: { type: integer }
        state:
          type: string
          enum: [accepted, quarantined, superseded]
        unitMinutes:
          type: [number, 'null']
        workforce:
          type: [number, 'null']
        quarantineReasons:
          type: array
          items: { type: string }
          description: Named defects; empty for an accepted revision.
      required: [sourceTaskKey, revision, state, unitMinutes, workforce, quarantineReasons]